# WIP for JCBC-1099 : N1ql query result parsing has poor performance for big document