@daschl 2017-08-08
I've been breeding over this change for a bit now and I think there might be a cleaner way with buffer slices - I'll let you know where I end up with this.