@daschl 2017-08-22
@benbenw we merged this in so feel free to close - thanks again! :)