@ingenthr 2017-08-28
Merged through Gerrit.  Thanks!